+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.0
+
+* Renamed `GitHubHookRequest` to `HookRequest`.
+
+* Added `IssuesHookRequest` subclass.
+
 ## 0.0.1
 
 - Getting started...